requirements.txt
# Core dependencies
# Requires Python 3.9 or higher

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# API
requests>=2.31.0
python-binance>=1.0.17

# Machine Learning - PyTorch
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Machine Learning - scikit-learn
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0

# Gradient Boosting
lightgbm>=4.1.0
xgboost>=2.0.0

# Technical indicators
ta>=0.11.0
pandas-ta
ta>=0.10.2>=0.3.14b

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Experiment tracking
wandb>=0.15.0
tensorboard>=2.13.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0
gym>=0.26.0
scipy>=1.11.0

# Dashboard
streamlit>=1.28.0

# Jupyter (optional, for notebooks)
jupyter>=1.0.0
ipykernel>=6.23.0

# Additional for complete system
catboost>=1.2.0

